test: move tpch decompile tests to separate file
