+++ ibis/expr/tests/snapshots/test_decompile_tpch/test_parse_sql_tpch/tpch3/out.tpch.py
